% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data.Metabric}
\alias{Data.Metabric}
\title{Gene expression profiling of METARBIC breast cancer dataset, 275 triple negative breast cancer samples are picked.}
\format{
a list of five components.
\itemize{
\item{Expression: }{Expression matrix with 275 samples(Rows) and 13964  genes(Cols)}
\item{OS: }{overall survival}
\item{OS.event: }{censoring indicators where 1 means events and 0 means censoring}
\item{covariate: }{covariates matrix: age, tumor size, number of lymph nodes and tumor grade}
\item{PAM50: }{PAM50 labels}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE47460}
}
\usage{
Data.Metabric
}
\description{
Gene expression profiling of METARBIC breast cancer dataset, 275 triple negative breast cancer samples are picked.
}
\keyword{datasets}
