% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Data.ScanB}
\alias{Data.ScanB}
\title{Gene expression profiling of SCAN-B breast cancer dataset, 151 triple negative breast cancer samples are picked.}
\format{
a list of five components.
\itemize{
\item{Expression: }{Expression matrix with 151 samples(Rows) and 18964  genes(Cols)}
\item{OS: }{overall survival}
\item{OS.event: }{censoring indicators where 1 means events and 0 means censoring}
\item{covariate: }{covariates matrix: age, tumor size, number of lymph nodes and tumor grade}
\item{PAM50: }{PAM50 labels}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE60789}
}
\usage{
Data.ScanB
}
\description{
Gene expression profiling of SCAN-B breast cancer dataset, 151 triple negative breast cancer samples are picked.
}
\keyword{datasets}
