% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.ogClust.R
\name{fit.ogClust}
\alias{fit.ogClust}
\title{Title Fit ogClust mixture model}
\usage{
fit.ogClust(
  n,
  K,
  np,
  NG,
  lambda,
  alpha,
  G,
  Y,
  X,
  theta_int,
  robust = "none",
  tau = 1.345
)
}
\arguments{
\item{n}{the number of samples}

\item{K}{the number of subgroups}

\item{np}{the number of}

\item{NG}{the number of genes}

\item{lambda}{the penalty parameter for sparsity}

\item{alpha}{the tuning parameter for elastic net}

\item{G}{the input gene expression matrix}

\item{Y}{the input vector of outcome}

\item{X}{the input vector of covariates}

\item{theta_int}{initial values of parameters}

\item{robust}{choose the type of robustness}

\item{tau}{set the cutoff for huber loss if robust is huberk}
}
\value{
An objects of class `ogClust`
\itemize{
 \item{'res'}{a vector of parameter estimates,likelihood `ll`, `R2`, `AIC`, `BIC` and tuning parameter `lambda`}
 \item{'prob'}{predicted probability for belonging to each subgroup}
 \item{'Y_prd'}{predicted outcome}
 \item{'grp_assign'}{prediced group assignement}
}
}
\description{
Title Fit ogClust mixture model
}
\details{
to be filled later
}
\examples{
  data(lung) #load lung dataset

  # extract gene expression G, covariate X, survival time Y
  G=lung$G
  X=lung$X
  Y=lung$Y

  # number of subjects
  n=nrow(G)
  # number of genes
  NG=ncol(G)
  # number of covariates
  np=ncol(X)
  # number of clusters
  K=3
  # tuning parameter
  lambda=0.13

  # set initial values
  beta_int = runif(np, 0, 3)
  gamma_int = runif((K - 1) * (NG + 1), 0, 1)
  beta0_int = runif(K, 0, 3)
  sigma2_int = runif(1, 1, 3)
  theta_int = c(beta_int, gamma_int, beta0_int, sigma2_int)

  # fit ogClust, robust is FALSE
  fit.res<-fit.ogClust(n=n, K=K, np=np, NG=NG, lambda=lambda,
                    alpha=0.5, G=G, Y=Y, X=X,theta_int=theta_int)
  # fit ogClust, robust method is median-truncation
  fit.res2<-fit.ogClust(n=n, K=K, np=np, NG=NG, lambda=lambda,
                     alpha=0.5, G=G, Y=Y, X=X,theta_int=theta_int,robust = 'median')
  # fit ogClust, robust method is Huber
  fit.res3<-fit.ogClust(n=n, K=K, np=np, NG=NG, lambda=lambda,
                     alpha=0.5, G=G, Y=Y, X=X,theta_int=theta_int, robust ='huber',tau=4.345)
  # fit ogClust, robust method is adaptive Huber
  fit.res4<-fit.ogClust(n=n, K=K, np=np, NG=NG, lambda=lambda,
                     alpha=0.5, G=G, Y=Y, X=X,theta_int=theta_int, robust='hubertf')
}
