% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.ogClust.surv.R
\name{fit.ogClust.surv}
\alias{fit.ogClust.surv}
\title{Title Fit ogClust mixture model}
\usage{
fit.ogClust.surv(
  n,
  K,
  np,
  NG,
  lambda,
  alpha,
  G,
  Y,
  X,
  delta,
  theta_int,
  dist = "loglogistic"
)
}
\arguments{
\item{n}{the number of samples}

\item{K}{the number of subgroups}

\item{np}{the number of prognostic covariates}

\item{NG}{the number of genes}

\item{lambda}{the penalty parameter for sparsity of genes}

\item{alpha}{the tuning parameter for the L2 loss of the penalty}

\item{G}{a matrix of gene expression, with subjects are rows and genes are columns}

\item{Y}{a vector of survival time}

\item{X}{a matrix of covariates, with subjects are rows and prognostic covariates are columns}

\item{delta}{a binary indicator of censoring, 0 means censored and 1 means event observed.}

\item{theta_int}{initial values of parameters}

\item{dist}{distribution of the survival time, defualt is loglogistic}
}
\value{
An object with class `ogClust`
\itemize{
 \item{'res'}{a vector of parameter estimates,likelihood `ll`, `R2`, `AIC`, `BIC` and tuning parameter `lambda`}
 \item{'prob'}{predicted probability for belonging to each subgroup}
 \item{'Y_prd'}{predicted outcome}
 \item{'grp_assign'}{prediced group assignement}
}
}
\description{
Title Fit ogClust mixture model
}
\details{
to be filled later
}
\examples{
  data('surv_dt') #load simulated survial data

  # extract gene expression G, covariate X, survival time Y
  # and censoring indicator delta
  G=surv_dt[,3:1002]
  X=surv_dt[,1:2]
  Y=surv_dt$time
  delta=surv_dt$event

  # number of subjects
  n=nrow(G)
  # number of genes
  NG=ncol(G)
  # number of covariates
  np=ncol(X)
  # number of clusters
  K=3
  # tuning parameter
  lambda=0.13

  # set initial values
  beta_int = runif(np, 0, 3)
  gamma_int = runif((K - 1) * (NG + 1), 0, 1)
  beta0_int = runif(K, 0, 3)
  sigma2_int = runif(1, 1, 3)
  theta_int = c(beta_int, gamma_int, beta0_int, sigma2_int)

  # fit ogClust
  fit.res<-fit.ogClust.surv(n=n, K=K, np=np, NG=NG, lambda=lambda,
                         alpha=0.5, G=G, Y=Y, X=X, delta, theta_int=theta_int)
}
