% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE47460_GPL14550}
\alias{GSE47460_GPL14550}
\title{Gene expression profiling of chronic lung disease for the Lung Genomics Research Consortium (GSE47460)
Microarray data from total RNA extracted from whole lung homogenate from subjects undergoing thoracic surgery.
There are two platforms in this GEO dataset and this data contains the platform GPL14550. We substract COPD and ILD patients in this dataset.}
\format{
a list of four components.
\itemize{
\item{outcome: }{outcome (FEV1) of 331 samples}
\item{Expression: }{Expression matrix with 331 samples(Cols) and 12958 genes(Rows)}
\item{Covariates: }{age and gender for 331 samples}
\item{label: }{Diagosis for 331 samples, either COPD or ILD}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE47460}
}
\usage{
GSE47460_GPL14550
}
\description{
Gene expression profiling of chronic lung disease for the Lung Genomics Research Consortium (GSE47460)
Microarray data from total RNA extracted from whole lung homogenate from subjects undergoing thoracic surgery.
There are two platforms in this GEO dataset and this data contains the platform GPL14550. We substract COPD and ILD patients in this dataset.
}
\keyword{datasets}
