% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogClust-package.R
\docType{package}
\name{ogClust-package}
\alias{ogClust}
\alias{ogClust-package}
\title{ogClust: Outcome guided clustering}
\description{
Outcome-guided clustering of high dimensional omic data(e.g. gene expression) to 
        identify the outcome assoicated subgroups toward precision medicine. Outcome could be 
        either continous or survival.
}
\author{
\strong{Maintainer}: Peng Liu \email{pel67@pitt.edu}

}
\keyword{internal}
