% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogClust_GM_LDA_surv.R
\name{ogClust_GM.LDA.Surv}
\alias{ogClust_GM.LDA.Surv}
\title{Title Fit ogClust_GM mixture model embedding the LDA for gene disease subtyping and with survival outcome}
\usage{
ogClust_GM.LDA.Surv(
  K,
  lambda,
  G,
  Y,
  X,
  delta,
  dist = "loglogistic",
  max_iter = 200
)
}
\arguments{
\item{K}{an integer defines the number of subgroups}

\item{lambda}{the regularization tuning parameter for sparsity}

\item{G}{the matrix for omics data. The rows are samples and columns are features.}

\item{Y}{the vector of time variable}

\item{X}{the vector of covariates}

\item{delta}{a binary indicator of censoring for time-to-event outcome. 0 means censored and 1 means event observed.}

\item{dist}{distribution of the survival time, defualt is \code{"loglogistic"}. The choices are \code{"weibull"},
\code{"exponential"}, \code{"gaussian"}, \code{"logistic"},\code{"lognormal"} and \code{"loglogistic"}.}

\item{max_iter}{the maximum number of iterations of EM algorithm. Default value is 200}
}
\value{
An object with class \code{"ogClust"}
\itemize{
\item{\code{par}}{ a list of parameter estimates}
\item{\code{lambda}}{lambda}
\item{\code{Y_prd}}{ predicted outcome}
\item{\code{grp_assign}}{ prediced group assignement}
}
}
\description{
Title Fit ogClust_GM mixture model embedding the LDA for gene disease subtyping and with survival outcome
}
\details{
The ogClust_GM.LDA.Surv is a unified latent generative model to perform clustering constructed
from omics data \code{G} with the guidance of outcome \code{Y}, and with covariate \code{X} to account for
the variability that is not related to subgrouping. A modified EM algorithm is applied for
numeric computation such that the liklihood is maximized. A posterior probability is obtain
for each subject belonging to each cluster.

Similar to ogClust_GM.Surv, ogClust_GM.LDA.Surv method performs feature selection, latent subtype characterization and outcome prediction simultaneously.
The difference is that ogClust_GM.LDA.Surv embeds the sparse linear discriminant analysis (LDA) to model gene disease subtyping.

Time variable \code{Y} and censoring indicator \code{delta} together defines a single time-to-event outcome.
We use accelerated-failure time(AFT) model to model time-to-event outcome, \code{dist} defines
the distribution of survival time.
}
\examples{
\dontrun{
  data('Data.Metabric') #load dataset

  # extract gene expression G, covariate X, outcome Y
  G=Data.Metabric$Expression
  X=as.matrix(Data.Metabric$covariate[,1])
  Y=Data.Metabric$OS
  delta<-Data.Metabric$OS.event
  Index.Mean<-order(apply(G,2,mean),decreasing = T)
  G <- G[,Index.Mean[1:(ncol(G)/2)]]
  Index.Sd<-order(apply(G,2,sd),decreasing = T)
  G<-G[,Index.Sd[1:(ncol(G)/2)]]
  G<-scale(G)
  # number of clusters
  K=2
  # tuning parameter
  lambda=0.007

  # fit ogClust
  fit.res<-ogClust_GM.LDA.Surv(K=K,lambda=lambda,delta = delta, G=G, Y=Y, X=X)
}
}
