% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogClust_WJL_surv.R
\name{ogClust_WJL.Surv}
\alias{ogClust_WJL.Surv}
\title{ogClust_WJL mehtod to cluster with survival outcome}
\usage{
ogClust_WJL.Surv(
  x,
  G,
  y,
  y.ind,
  c_center = NULL,
  lambda,
  v_int = NULL,
  pi_int = NULL,
  K = 3,
  max_iter = 200,
  w_outcome = 0.5,
  w_G = 0.5,
  z_int = NULL
)
}
\arguments{
\item{x}{covariates in the outcome association model (e.g., age, gender, BMI),  n*q matrix where n is sample size and q is number of covariates}

\item{G}{Gene expression data (p*n matrix), where each row is a gene and each column is a sample}

\item{y}{outcome, a vector of length n}

\item{y.ind}{the censor indicator for the outcome. 1 means events and 0 means censor.}

\item{c_center}{a n*K matrix which determines the initial clusters.}

\item{lambda}{a tuning parameter determines gene selection.}

\item{v_int}{initial variance for each gene, optional}

\item{pi_int}{initial probability for each cluster, optional}

\item{K}{Number of Clusters}

\item{max_iter}{maximum number of iterations}

\item{w_outcome}{weight for outcome}

\item{w_G}{weight for gene expression}

\item{z_int}{a initial cluster assignment matrix. A n*K matrix where each row denotes the probability
for a sample belongs to each cluster. If z_int is given, c_center will be ignored.}
}
\value{
result_list: A list with final estimators.
\itemize{
\item{mu: }{p\emph{K matrix where each row represents the cluster centers for each feature}
\item{sigma: }{a vector of length p, the standard deviation for each feature}
\item{z: }{n}K matrix where each row represents the probability the sample belongs to each cluster}
\item{pi: }{a vector of length K, mixed probability for each cluster}
\item{int_coef: }{a K+q vector, the first K elements are the intercept estimated for each cluster. The next q values are the common coefficient for the covariates.}
\item{int_sigma_coef: }{standard deviation estimation in the outcome association model.}
}
Other three outputs:
\itemize{
\item{BIC: }{BIC from the final model}
\item{lik: }{optimal log likelihood achieved}
\item{iter: }{Number of iterations before convergence}
}
}
\description{
ogClust_WJL mehtod to cluster with survival outcome
}
