% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region_lambda_surv.R
\name{region_lambda_surv}
\alias{region_lambda_surv}
\title{efficient lambda grids for a given K, for survival outcome}
\usage{
region_lambda_surv(
  lambda1 = 18,
  lambda2 = 0,
  iteration = 10,
  Y,
  G,
  X,
  center,
  w,
  K = K,
  delta
)
}
\arguments{
\item{lambda1}{a large lambda which gives close to 0 informative genes}

\item{lambda2}{a small lambda which includes almost all the genes (default is lambda2=0)}

\item{iteration}{Number of iterations}

\item{Y}{outcome}

\item{G}{gene expression matrix (row is samples, column is genes)}

\item{X}{covariates matrix}

\item{center}{initial center for ogClust}

\item{w}{weight of outcome}

\item{K}{Number of clusters}

\item{delta}{censor indicator of survival outcome}
}
\value{
a list with two components
\itemize{
\item{lambda: }{generated lambda grids}
\item{num: }{number of informative genes corresponding to each lambda}
}
}
\description{
efficient lambda grids for a given K, for survival outcome
}
