% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rsquare.R
\name{Rsquare}
\alias{Rsquare}
\title{Get the Rsquare value for gene expression and outcome (continuous)}
\usage{
Rsquare(cluster, Y, G, X)
}
\arguments{
\item{cluster}{cluster assignment}

\item{Y}{outcome}

\item{G}{Gene expression data (p*n matrix), where each row is a gene and each column is a sample}

\item{X}{covariates matrix}
}
\value{
a list with two components
\itemize{
\item{R2.outcome: }{R2 of outcome}
\item{R2.gene: }{R2 of p genes}
}
}
\description{
Get the Rsquare value for gene expression and outcome (continuous)
}
\examples{
\dontrun{
#mod is the model from ogClust
cluster<-apply(mod$result_list$z,1,which.max)
#y is continuous outcome, G is gene expression matrix, x is covariates matrix.
Rsquare(cluster,Y = y,G = G,X = x)
}
}
