% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_surv_dt.R
\docType{data}
\name{surv_dt}
\alias{surv_dt}
\title{Simulated survival data}
\format{
a list containing gene expression matrix `G`,
covariate matrix `X` and survival time `Y` and a binary
indicator `delta` for censoring.
}
\usage{
data(surv_dt)
}
\description{
A simulated survival data, with expression of 1000
genes, two prognostic covariates, survival time,
plus an indicator for censoring.
}
\examples{
  data('surv_dt') #load simulated survial data

  # extract gene expression G, covariate X, survival time Y
  # and censoring indicator delta
  G=surv_dt[,3:1002]
  X=surv_dt[,1:2]
  Y=surv_dt$time
  delta=surv_dt$event

  # number of subjects
  n=nrow(G)
  # number of genes
  NG=ncol(G)
  # number of covariates
  np=ncol(X)
  # number of clusters
  K=3
  # tuning parameter
  lambda=0.13

  # set initial values
  beta_int = runif(np, 0, 3)
  gamma_int = runif((K - 1) * (NG + 1), 0, 1)
  beta0_int = runif(K, 0, 3)
  sigma2_int = runif(1, 1, 3)
  theta_int = c(beta_int, gamma_int, beta0_int, sigma2_int)

  # fit ogClust
  fit.res<-fit.ogClust.surv(n=n, K=K, np=np, NG=NG, lambda=lambda,
                         alpha=0.5, G=G, Y=Y, X=X, delta, theta_int=theta_int)
}
\author{
Peng Liu, 2020-06-21
}
\keyword{datasets}
