% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.ogClust.surv.R
\name{fit.ogClust.surv}
\alias{fit.ogClust.surv}
\title{Title Fit ogClust mixture model}
\usage{
fit.ogClust.surv(
  n,
  K,
  np,
  NG,
  lambda,
  alpha,
  G,
  Y,
  X,
  delta,
  theta_int,
  dist = "loglogistic"
)
}
\arguments{
\item{n}{an integer defines number of samples}

\item{K}{an integer defines the number of subgroups}

\item{np}{an integer defines the number of covariates}

\item{NG}{an integer defines the number of features of omics data}

\item{lambda}{the regularization tuning parameter for sparsity}

\item{alpha}{the L2 regularization tuning parameter}

\item{G}{the matrix for omics data. The rows are samples and columns are features.}

\item{Y}{the vector of time variable}

\item{X}{the vector of covariates}

\item{delta}{a binary indicator of censoring for time-to-event outcome. 0 means censored and 1 means event observed.}

\item{theta_int}{a vector of initial values for parameters to estimate}

\item{dist}{distribution of the survival time, defualt is \code{"loglogistic"}. The choices are \code{"weibull"},
\code{"exponential"}, \code{"gaussian"}, \code{"logistic"},\code{"lognormal"} and \code{"loglogistic"}.}
}
\value{
An object with class \code{"ogClust"}
\itemize{
 \item{\code{res}}{ a vector of parameter estimates,likelihood \code{"ll"}, \code{"R2"}, \code{"AIC"}, \code{"BIC"} and tuning parameter \code{lambda}}
 \item{\code{prob}}{ predicted probability for belonging to each subgroup}
 \item{\code{Y_prd}}{ predicted outcome}
 \item{\code{grp_assign}}{ prediced group assignement}
}
}
\description{
Title Fit ogClust mixture model
}
\details{
The ogClust is a unified latent generative model to perform clustering constructed
from omics data \code{G} with the guidance of outcome \code{Y}, and with covariate \code{X} to account for
the variability that is not related to subgrouping. A modified EM algorithm is applied for
numeric computation such that the liklihood is maximized. A posterior probability is obtain
for each subject belonging to each cluster.

ogClust method performs feature selection, latent subtype characterization and outcome prediction simultaneously.
We use either LASSO penalty or LASSO penalty plus L2 regularization( \eqn{lambda*(L1 + alpha*L2)})
Parameter \code{lambda} is the penalty tuning parameter, and \code{alpha} tunes the L2 regularization.
Time variable \code{Y} and censoring indicator \code{delta} together defines a single time-to-event outcome.
We use accelerated-failure time(AFT) model to model time-to-event outcome, \code{dist} defines
the distribution of survival time.
}
\examples{
  data('surv_dt') #load simulated survial data

  # extract gene expression G, covariate X, survival time Y
  # and censoring indicator delta
  G=surv_dt[,3:1002]
  X=surv_dt[,1:2]
  Y=surv_dt$time
  delta=surv_dt$event

  # number of subjects
  n=nrow(G)
  # number of genes
  NG=ncol(G)
  # number of covariates
  np=ncol(X)
  # number of clusters
  K=3
  # tuning parameter
  lambda=0.13

  # set initial values
  beta_int = runif(np, 0, 3)
  gamma_int = runif((K - 1) * (NG + 1), 0, 1)
  beta0_int = runif(K, 0, 3)
  sigma2_int = runif(1, 1, 3)
  theta_int = c(beta_int, gamma_int, beta0_int, sigma2_int)

  # fit ogClust
  fit.res<-fit.ogClust.surv(n=n, K=K, np=np, NG=NG, lambda=lambda,
                         alpha=0.5, G=G, Y=Y, X=X, delta, theta_int=theta_int)
}
